# Implement multiple dispatch (#26468)