--- a/aten/src/ATen/native/BinaryOps.cpp
+++ b/aten/src/ATen/native/BinaryOps.cpp
@@ -20,16 +20,6 @@ static constexpr char alpha_mismatch_err[] =
   "For integral input tensors, argument alpha must not be a floating point number.";
 
 Tensor& add_out(Tensor& result, const Tensor& self, const Tensor& other, Scalar alpha) {
-  if (other.is_sparse()) {
-    if (self.is_sparse()) {
-      at::_sparse_add_out(result, self, other, alpha);
-    } else {
-      at::_sparse_dense_add_out(result, self, other, alpha);
-    }
-    return result;
-  } else if (self.is_sparse()) {
-    AT_ERROR("add(sparse, dense) is not supported. Use add(dense, sparse) instead.");
-  }
   auto iter = TensorIterator::binary_op(result, self, other,
     /*check_mem_overlap=*/true);
   TORCH_CHECK(! alpha.isBoolean() || iter.dtype() == ScalarType::Bool, "Boolean alpha only supported for boolean results");
@@ -41,10 +31,6 @@ Tensor& add_out(Tensor& result, const Tensor& self, const Tensor& other, Scalar 
 
 Tensor add(const Tensor& self, const Tensor& other, Scalar alpha) {
   Tensor result;
-  if (other.is_sparse()) {
-    result = at::empty({0}, self.options());
-    return native::add_out(result, self, other, alpha);
-  }
   auto iter = TensorIterator::binary_op(result, self, other);
   TORCH_CHECK(! alpha.isBoolean() || iter.dtype() == ScalarType::Bool, "Boolean alpha only supported for boolean results");
   TORCH_CHECK(isFloatingType(iter.dtype()) || alpha.isIntegral(true), alpha_mismatch_err);
@@ -57,13 +43,6 @@ Tensor& add_(Tensor& self, const Tensor& other, Scalar alpha) {
 }
 
 Tensor& div_out(Tensor& result, const Tensor& self, const Tensor& other) {
-  if (self.is_sparse()) {
-    if (other.dim() != 0) {
-      AT_ERROR("div(): sparse division only supports division by a scalar ",
-        "(got shape ", other.sizes(), " for argument 'other')");
-    }
-    return at::_sparse_div_zerodim_out(result, self, other);
-  }
   auto iter = TensorIterator::binary_op(result, self, other,
     /*check_mem_overlap=*/true);
   div_stub(iter.device_type(), iter);
@@ -72,10 +51,6 @@ Tensor& div_out(Tensor& result, const Tensor& self, const Tensor& other) {
 
 Tensor div(const Tensor& self, const Tensor& other) {
   Tensor result;
-  if (self.is_sparse()) {
-    result = at::empty({0}, self.options());
-    return native::div_out(result, self, other);
-  }
   auto iter = TensorIterator::binary_op(result, self, other);
   div_stub(iter.device_type(), iter);
   return iter.output();
@@ -86,9 +61,6 @@ Tensor& div_(Tensor& self, const Tensor& other) {
 }
 
 Tensor& mul_out(Tensor& result, const Tensor& self, const Tensor& other) {
-  if (self.is_sparse() || other.is_sparse()) {
-    return at::_sparse_mul_out(result, self, other);
-  }
   auto iter = TensorIterator::binary_op(result, self, other,
     /*check_mem_overlap=*/true);
   mul_stub(iter.device_type(), iter);
@@ -97,10 +69,6 @@ Tensor& mul_out(Tensor& result, const Tensor& self, const Tensor& other) {
 
 Tensor mul(const Tensor& self, const Tensor& other) {
   Tensor result;
-  if (self.is_sparse() || other.is_sparse()) {
-    result = at::empty({0}, self.options());
-    return native::mul_out(result, self, other);
-  }
   auto iter = TensorIterator::binary_op(result, self, other);
   mul_stub(iter.device_type(), iter);
   return iter.output();
@@ -122,19 +90,6 @@ static inline void sub_check(const Tensor& self, const Tensor& other) {
 
 Tensor& sub_out(Tensor& result, const Tensor& self, const Tensor& other, Scalar alpha) {
   sub_check(self, other);
-  if (other.is_sparse()) {
-    if (!self.sizes().equals(other.sizes())) {
-      AT_ERROR("sizes do not match");
-    }
-    if (self.is_sparse()) {
-      at::_sparse_add_out(result, self, other, -alpha);
-    } else {
-      at::_sparse_dense_add_out(result, self, other, -alpha);
-    }
-    return result;
-  } else if (self.is_sparse()) {
-    AT_ERROR("sub(sparse, dense) is not supported. Use sub(dense, sparse) instead.");
-  }
   auto iter = TensorIterator::binary_op(result, self, other,
     /*check_mem_overlap=*/true);
   TORCH_CHECK(isFloatingType(iter.dtype()) || alpha.isIntegral(false), alpha_mismatch_err);
@@ -146,10 +101,6 @@ Tensor& sub_out(Tensor& result, const Tensor& self, const Tensor& other, Scalar 
 Tensor sub(const Tensor& self, const Tensor& other, Scalar alpha) {
   sub_check(self, other);
   Tensor result;
-  if (other.is_sparse()) {
-    result = at::empty({0}, self.options());
-    return native::sub_out(result, self, other, alpha);
-  }
   auto iter = TensorIterator::binary_op(result, self, other);
   TORCH_CHECK(isFloatingType(iter.dtype()) || alpha.isIntegral(false), alpha_mismatch_err);
   sub_stub(iter.device_type(), iter, alpha);
@@ -197,12 +148,18 @@ Tensor& add_(Tensor& self, Scalar other, Scalar alpha) {
   return native::add_(self, wrapped_scalar_tensor(other), alpha);
 }
 
+// WARNING: There doesn't appear to be any testing for this function
+// with sparse self input.
 Tensor div(const Tensor& self, Scalar other) {
-  return native::div(self, wrapped_scalar_tensor(other));
+  return self.div(wrapped_scalar_tensor(other)); // redispatch!
 }
 
+// WARNING: This function, with a sparse self, is currently only
+// exercised by DistributedDataParallelTest.test_sparse_gradients
+// (you need to exercise it from C++, because this overload is never
+// used for Python)
 Tensor& div_(Tensor& self, Scalar other) {
-  return native::div_(self, wrapped_scalar_tensor(other));
+  return self.div_(wrapped_scalar_tensor(other)); // redispatch!
 }
 
 Tensor mul(const Tensor& self, Scalar other) {