… aten/src/ATen/core/OpsAlreadyMovedToC10.cpp
… aten/src/ATen/core/Variadic.h
… aten/src/ATen/native/LegacyBridge.cpp
… aten/src/ATen/native/native_functions.yaml
… aten/src/ATen/native/sparse/SparseTensorMath.h
… aten/src/ATen/native/sparse/cuda/SparseCUDATensor.cpp
… aten/src/ATen/native/sparse/cuda/SparseCUDATensorMath.cu
… aten/src/ATen/templates/TensorBody.h
… aten/src/ATen/templates/TensorMethods.h
… c10/core/TensorTypeId.h
… test/test_nn.py
… test/test_sparse.py
… test/test_torch.py
… tools/autograd/templates/Functions.cpp
… torch/csrc/utils/variadic.h
… torch/lib/c10d/ProcessGroupGloo.cpp